requirements.txt
asgiref==3.4.1
Django==3.2.11
djangorestframework==3.13.1
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
